TRAVIS rsync & cache /var/log on build failure
When a build fails on travis, we often need to inspect cluster node
logs. With these edits, we will be able to persist the logs for
inspection. First, we check to see if the build failed. If it did fail,
we use rsync in archive mode to pull down the /var/log directory from
the first control node, put it in .mantl-ci-log/<build number>, and have
travis cache that directory. The location of the cache is
https://travis-ci.org/CiscoCloud/mantl/caches