TRAVIS install rsync to image, get cloud-ini logs
We need rsync to be installed on the Docker image in order to run the
commands to get the logs. I also limit the logs to the cloud-init logs
that we need for the DO support team.

Additionally, I added a few dev packages to the image, so that the
python dependencies can compile